change code to big int

+++ database/migrations/2022_12_10_053421_create_binaries_table.php
@@ -17,7 +17,7 @@ class CreateBinariesTable extends Migration
         Schema::create('binaries', function (Blueprint $table) {
             $table->id();
             $table->string('game');
-            $table->unsignedInteger('code')->nullable();
+            $table->unsignedBigInteger('code')->nullable();
             $table->string('file');
             $table->string('target');
             $table->string('version');
